first couple of cases handled for clumper. state reader removed for now, as all tests are deprecated in favor of new clumping logic
