+++ .gitignore
@@ -1,1 +1,2 @@
 .rvmrc
+*DS_Store*
