--- players/seek_and_destroy/ship.rb
@@ -1,4 +1,0 @@
-require_relative "./ship.rb"
-class Ship
-  attr_accessor :orientation
-end
